+++ lib/research_metadata.rb
@@ -1,7 +1,7 @@
 require 'puree'
 require 'datacite/mapping'
 require 'research_metadata/transformer/dataset'
-require 'research_metadata/transformer/publication'
+require 'research_metadata/transformer/research_output'
 require 'research_metadata/transformer/thesis'
 require 'research_metadata/version'