+++ lib/research_metadata/transformer/research_output.rb
@@ -2,30 +2,31 @@ module ResearchMetadata
 
   module Transformer
 
-    # Extracts publication metadata from the Pure Research Information System
+    # Extracts research output metadata from the Pure Research Information System
     # and converts it into the DataCite format.
     # For text-based resources.
+    # @note Do not use directly.
     #
-    class Publication
+    class ResearchOutput
 
       # @param config [Hash]
-      # @option config [String] :url The URL of the Pure host.
-      # @option config [String] :username The username of the Pure host account.
-      # @option config [String] :password The password of the Pure host account.
+      # @option config [String] :url URL of the Pure host
+      # @option config [String] :username Username of the Pure host account
+      # @option config [String] :password Password of the Pure host account
+      # @option config [String] :api_key API key of the Pure host account
       def initialize(config)
         @config = config
-        @publication_extractor = Puree::Extractor::Publication.new config
+        @research_output_extractor = Puree::Extractor::ResearchOutput.new config
       end
 
-      # Publication transformation
+      # Research output transformation
       #
       # @param id [String]
-      # @param uuid [String]
       # @param doi [String]
       # @return [String, nil]
-      def transform(id: nil, uuid: nil, doi: nil)
-        @publication = extract uuid: uuid, id: id
-        return nil if !@publication
+      def transform(id:, doi:)
+        @research_output = @research_output_extractor.find id
+        return nil if !@research_output
         person_o = person
         file_o = file
         resource = ::Datacite::Mapping::Resource.new(
@@ -75,7 +76,7 @@ module ResearchMetadata
       end
 
       def description
-        desc = @publication.description
+        desc = @research_output.description
         if desc
           d = ::Datacite::Mapping::Description.new value: desc,
                                              type: ::Datacite::Mapping::DescriptionType::ABSTRACT
@@ -85,16 +86,8 @@ module ResearchMetadata
         end
       end
 
-      def extract(uuid: nil, id: nil)
-        if !uuid.nil?
-          return @publication_extractor.find uuid: uuid
-        else
-          return @publication_extractor.find id: id
-        end
-      end
-
       def file
-        @publication.files
+        @research_output.files
       end
 
       def identifier(doi)
@@ -102,7 +95,7 @@ module ResearchMetadata
       end
 
       def language
-        @publication.locale
+        @research_output.language
       end
 
       def name_identifier_orcid(person)
@@ -120,9 +113,9 @@ module ResearchMetadata
         o['creator'] = []
         o['contributor'] = []
         all_persons = []
-        all_persons << @publication.persons_internal
-        all_persons << @publication.persons_external
-        all_persons << @publication.persons_other
+        all_persons << @research_output.persons_internal
+        all_persons << @research_output.persons_external
+        all_persons << @research_output.persons_other
         all_persons.each do |person_type|
           person_type.each do |individual|
             pure_role =individual.role.gsub(/\s+/, '')
@@ -142,7 +135,7 @@ module ResearchMetadata
             if human
               if individual.uuid
                 person_extractor = Puree::Extractor::Person.new @config
-                person = person_extractor.find uuid: individual.uuid
+                person = person_extractor.find individual.uuid
                 if person
                   identifier = name_identifier_orcid person
                   human.identifier = identifier if identifier
@@ -163,16 +156,12 @@ module ResearchMetadata
       end
 
       def publication_year
-        @publication.statuses.each do |i|
+        @research_output.publication_statuses.each do |i|
           if i.stage === 'Published' || i.stage === 'Unpublished'
             return i.date.year
           end
         end
         nil
-      end
-
-      def publisher
-        @publication.publisher || 'Publisher unspecified'
       end
 
       def resource_type
@@ -183,17 +172,22 @@ module ResearchMetadata
       end
 
       def subjects
-        @publication.keywords.map { |i| ::Datacite::Mapping::Subject.new value: i }
+        @research_output.keywords.map { |i| ::Datacite::Mapping::Subject.new value: i }
       end
 
       def titles
         arr = []
-        title = ::Datacite::Mapping::Title.new value: @publication.title
+        title = ::Datacite::Mapping::Title.new value: @research_output.title
         arr << title
-        subtitle = @publication.subtitle
+        subtitle = @research_output.subtitle
         if subtitle
           arr << ::Datacite::Mapping::Title.new(value: subtitle,
                                                 type: ::Datacite::Mapping::TitleType::SUBTITLE)
+        end
+        translated_title = @research_output.translated_title
+        if translated_title
+          arr << ::Datacite::Mapping::Title.new(value: translated_title,
+                                                type: ::Datacite::Mapping::TitleType::TRANSLATED_TITLE)
         end
         arr
       end