+++ spec/spec_helper.rb
@@ -1,4 +1,0 @@
-require 'puree'
-require 'datacite/mapping'
-require 'research_metadata/transformer/dataset'
-require 'research_metadata/version'