+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 2.0.0 2017-12-20
+### Added
+- Research output - translated title.
+- Testing against known live data.
+
+### Changed
+- For Pure API 59.
 
 ## 1.1.2 2017-07-12
 ### Fixed